% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biaslayer.R
\name{biaslayer}
\alias{biaslayer}
\title{biaslayer: Function to create a bias layer}
\usage{
biaslayer(occs_df, longitude, latitude, raster_mold)
}
\arguments{
\item{occs_df}{A data.frame with the occurrence data and coordinates.}

\item{longitude}{A character vector of the column name of longitude.}

\item{latitude}{A character vector of the column name of latitude.}

\item{raster_mold}{A raster of the calibration area. This will serve just as a mold to make the bias layer.}
}
\description{
biaslayer estimates a raster that represents the sampling bias; it can be used in the \code{\link{maxent_call}{ntbox}} function.
}
\details{
The biaslayer uses the 2 dimensional kernel desity estimator \code{\link{kde2d}{MASS}}.
}
\examples{
# Birds of North America
birds <- read.csv(system.file("extdata",
                              "birdsN_America.csv",
                              package = "ntbox"))
bio01 <- raster::raster(system.file("extdata",
                                    "bio01_bias.tif",
                                    package = "ntbox"))

biasBirds <- biaslayer(occs_df = birds,
                       longitude = "Longitude",
                       latitude = "Latitude",
                       raster_mold = bio01)

raster::plot(biasBirds)

}
