% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_model.R
\name{bin_model}
\alias{bin_model}
\title{bin_model: Binarize a model using a threshold}
\usage{
bin_model(model, occs, percent)
}
\arguments{
\item{model}{A continuos raster model of suitability values}

\item{occs}{Ocurrence data with two colomuns (longitude and latitude).}

\item{percent}{Type of thresholding method. Values go from 0-100 percent
of the data. 0 is equivalent to the minimum training presence.}
}
\value{
A binary map of the prediction.
}
\description{
Function to binarize a continuos model given a threshold
}
\examples{
\dontrun{

model_p <- system.file("extdata/ambystoma_model.tif",
                       package = "ntbox")
model <- raster::raster(model_p)
data_p <- system.file("extdata/ambystoma_validation.csv",
                      package = "ntbox")
data <- read.csv(data_p)
occs <- data[which(data$presence_absence==1),]
binary <- bin_model(model,occs,percent = 5)
raster::plot(binary)
}
}
