% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dup.R
\name{clean_dup}
\alias{clean_dup}
\title{Function to clean duplicated longitude and latitude data}
\usage{
clean_dup(data, longitude, latitude, threshold = 0)
}
\arguments{
\item{data}{A data.frame with longitude and latitude data}

\item{longitude}{A character vector of the column name of longitude.}

\item{latitude}{A character vector of the column name of latitude.}

\item{threshold}{A numerc value representig the distance (in grades) between coordinates
to be considered as a duplicate.}
}
\value{
Returns a data.frame with coordinate data from species
}
\description{
Clean duplicated longitude and latitude data using threshold distance
             which is a distance (in grades) between points to be considered
             duplicates.
}
\examples{
# Species genus
genus <- "ambystoma"
# Species name
species <- "tigrinum"
# GBIF search
# ambystoma_tigrinum <- searh_gbif_data(genus,species,
#                                        occlim=100,
#                                        writeFile=FALSE)
# ambystoma_tigrinum_clean <- clean_dup(ambystoma_tigrinum,
#                                     longitude = "longitude",
#                                     latitude = "latitude",
#                                     threshold=0.0)
# Check the dimensions of  data
# dim(ambystoma_tigrinum)
# dim(ambystoma_tigrinum_clean)
}

