% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confu_mat_optim.R
\name{confu_mat_optim}
\alias{confu_mat_optim}
\title{Function to find the cut-off threshold that optimizes the confusion matrix}
\usage{
confu_mat_optim(sdm_raster, valData, longitude, latitude, pres_aus,
  optim_by = "kappa", th_range = c(0, 1), step = 0.005)
}
\arguments{
\item{sdm_raster}{SDM predicition raster}

\item{valData}{A data.frame or a matrix with validation data. The data must contain
ongitude, latitude and a column of presences (1) and absences (0).}

\item{longitude}{Column name of the longitude data}

\item{latitude}{Column name of the latitude data}

\item{pres_aus}{Column name of presences and absences. The presences have to represented by
1 and the absences with 0.}

\item{optim_by}{The optimation should be by kappa (Kappa)  metric or by tss (TSS) metric.}

\item{th_range}{A vector with the minimum and max range of threshold values used in the searching process.}

\item{step}{A numeric value representing the step size to be used in the searching process.}
}
\value{
Returns a data.frame with the values of: confusion matrix, kappa, tss, prevalence,
specificity, sensibility,correct classification rate (correct_class_rate), miss classification rate (miss_cla_rate),
positive predictive power (posit_pre_pow), negative predictive power (nega_pre_pow), comission error (comission_error)
omission error (omission_error).
}
\description{
The function finds out which is the best cut-off threshold that
optimizes the confusion matrix based on Kappa or TSS
}

