% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_finder.R
\name{correlation_finder}
\alias{correlation_finder}
\title{Function to find out strong correlations in a correlation matrix}
\usage{
correlation_finder(cor_mat, threshold, verbose = TRUE)
}
\arguments{
\item{cor_mat}{A correlation matrix.}

\item{threshold}{Threshold valuefrom which it is considered that the correlation is high.}

\item{verbose}{Verbose output.}
}
\value{
Returns a vector with variable names that can summarize the information.
}
\description{
The function finds out which variables have strong
correlations according to a correlation threshold. The output
returns a list of variables names that can summarize the information
and removes the variables that are redundant.
}
\examples{
# library(Matrix)
# set.seed(111)
# mat_cor <-Matrix(runif(16), 4)
# mat_cor <- forceSymmetric(mat_cor)
# diag(mat_cor) <- 1
# colnames(mat_cor) <- c(paste0("var_",1:4))
# Correlation matrix
# cor_mat <- data.frame(as.matrix(mat_cor))
# correlation_finder(cor_mat = cor_mat ,threshold = 0.7,verbose = TRUE)
}
