% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoidfit2.R
\name{ellipsoidfit2}
\alias{ellipsoidfit2}
\title{ellipsoidfit2: function to fit an ellipsoid model}
\usage{
ellipsoidfit2(
  envlayers,
  centroid,
  covar,
  level = 0.95,
  plot = T,
  size,
  xlab1 = "niche var 1",
  ylab1 = "niche var 2",
  zlab1 = "S",
  ...
)
}
\arguments{
\item{envlayers}{A RasterStack or RasterBrick object of the niche variables.}

\item{centroid}{A vector with the values of the centers of the ellipsoid (see \code{\link[ntbox]{cov_center}}).}

\item{covar}{The shape matrix (covariance) of the ellipoid (see \code{\link[ntbox]{cov_center}}).}

\item{level}{The proportion of points  to be included inside the ellipsoid}

\item{plot}{Logical If True a plot of the niche will be shown.}

\item{size}{The size of the points of the niche plot.}

\item{xlab1}{For x label for 2-dimensional histogram}

\item{ylab1}{For y label for 2-dimensional histogram}

\item{zlab1}{For z label for 2-dimensional histogram}

\item{...}{Arguments passed to \code{\link[rgl]{plot3d}} function from rgl}
}
\value{
Returns a list containing a data.frame with the suitability values; a suitability raster;
a data.frame with the mahalanobis and euclidean distances to the centroid.
}
\description{
Function to fit an ellipsoid model using the shape matrix (covariance matrix)
of the niche variables. This is an optimized version of the \code{\link[ntbox]{ellipsoidfit}} function; the diffrence is that it does not give the table of distances to niche centroid.
}
\examples{
## Load niche data
\dontrun{
d_cardon <-  read.csv(system.file("extdata", "cardon_virtual.csv", package = "ntbox"))
## Compute the centroid and shape (covariance matrix) of the ellipsoid model.
covar_centroid <- cov_center(d_cardon,mve=TRUE,level=0.99,vars=c(3,4,5))
## RasterStack with the niche variables
nicheStack <- raster::stack(list.files(system.file("extdata",
                                       package = "ntbox"),
                                       pattern = ".asc$",
                                       full.names = TRUE))
# Fitting the ellipsoid model
 ellipsoidMod <- ellipsoidfit2(envlayers = nicheStack,
                               centroid = covar_centroid$centroid,
                               covar = covar_centroid$covariance,
                               level=0.99,
                               plot=TRUE,
                               size=3)
 raster::plot(ellipsoidMod)
}
}
