% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bio_oracle.R
\name{get_bio_oracle}
\alias{get_bio_oracle}
\title{get_bio_oracle: Get environmental layers from Bio-Oracle.}
\usage{
get_bio_oracle(
  period,
  var_type,
  model = NULL,
  scenario = NULL,
  sv_dir = getwd(),
  load2r = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{period}{Time period. Posible values are: "current","2050","2100","2200".}

\item{var_type}{Type of variable.  Posible values are: 'salinity','sea_surface_temperature',
'current_velocity','sea_water_temperature','sea_water_salinity','sea_ice_thickness',
'chlorophyll_concentration','sea_surface_salinity'.}

\item{model}{Climate model. Posible values are:"UKMO-HadCM3" and "AOGCM"}

\item{scenario}{Climate change scenario. Posible values are "a1b","a2","b1","rcp26","rcp45","rcp60","rcp85".}

\item{sv_dir}{Path to the directory where the layers will be saved. Default is the working directory of R session.}

\item{load2r}{Logical. Load layers into R?}

\item{parallel}{Download layers in parallel.}
}
\description{
Get bioclimatic layers from Bio-Oracle for current and future scenarios
}
\details{
For more details visit \url{http://www.bio-oracle.org/index.php}
}
\examples{
\dontrun{
swater_temp <- get_bio_oracle(period = "current",
                              var_type =  'sea_water_temperature',
                              model = NULL,
                              scenario = NULL,
                              sv_dir="~/Desktop/",
                              load2r = TRUE,
                              parallel = TRUE)
swater_temp_2100_AOGCM_rcp85 <- get_bio_oracle(period = "2100",
                                               var_type ='sea_water_temperature',
                                               model = "AOGCM",
                                               scenario = "rcp85",
                                               sv_dir="~/Desktop/",
                                               load2r = TRUE,
                                               parallel = TRUE)
}
}
\references{
Assis, J., Tyberghein, L., Bosh, S., Verbruggen, H., Serrao, E. A., & De Clerck, O. (2017). Bio-ORACLE v2.0: Extending marine data layers for bioclimatic modelling. Global Ecology and Biogeography.
}
\seealso{
\code{\link[ntbox]{get_envirem_elev}}, \code{\link[ntbox]{get_envirem_clim}},  \code{\link[ntbox]{get_chelsa}}
}
