% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bio_oracle_v2.R
\name{get_bio_oracle_v2}
\alias{get_bio_oracle_v2}
\title{Get BioOracle v2 Data Layers}
\usage{
get_bio_oracle_v2(
  period = "Present",
  rcp = NULL,
  depth = "Benthic.Min",
  variable = "Temperature",
  stat = "all",
  format = "tif",
  save_path = "."
)
}
\arguments{
\item{period}{A character string specifying the time period for which to retrieve the data. Options are "Present", "2050", or "2100". The default is "Present".}

\item{rcp}{A character string or vector of characters specifying the greenhouse gas emission scenarios to consider.
Can be one or more of: "RCP26", "RCP45", "RCP60", "RCP85". The default is \code{NULL}. It will only be considered for future periods (2050 and 2100).}

\item{depth}{A character string specifying the depth level for the data. Options are: "Benthic.Min", "Benthic.Max", "Benthic.Mean", "Surface". The default is "Benthic.Min".}

\item{variable}{A character string or vector of characters specifying the variable to retrieve data for. Options include:
"Temperature", "Salinity", "Primary.productivity", "pH", "Iron", "Ice.thickness", "Ice.cover", "Current.Velocity",
"Nitrate", "Phosphate", "Silicate", "Dissolved.oxygen", "Chlorophyll", "Phytoplankton", "Calcite", "Par",
"Diffuse.attenuation", "Cloud.cover", "Sea". The default is "Temperature".}

\item{stat}{A character string or vector of characters specifying the statistics to calculate for the selected variable. Options are:
'Min', 'Max', 'Mean', 'Lt.max', 'Lt.min', 'Range'. The default is "all", which selects all statistics.}

\item{format}{A character string specifying the file format for saving the data. Options are "tif" or "asc". The default is "tif".}

\item{save_path}{A character string specifying the directory where the downloaded and processed files will be saved. The default is the current directory ("./").}
}
\value{
A \code{SpatRaster} object from the \code{terra} package containing the processed and downloaded data layers.
If no data is found for the selected parameters, the function will stop and display an error message.
}
\description{
This function retrieves environmental and climatic data layers from BioOracle v2 based on different combinations of parameters
such as the time period, RCP (emission scenarios), depth, variable, and statistics. Data can be downloaded in \code{tif} or \code{asc} format
and saved in the specified directory.
}
\details{
This function filters the BioOracle v2 data based on the selected parameters and downloads the corresponding layers.
If the specified \code{save_path} does not exist, the function will create the necessary folders to store the files.
The downloaded files are unzipped, renamed according to the query parameters, and loaded into a \code{SpatRaster} object for further analysis.
}
\examples{
\dontrun{
# Download data for the present period and the variable 'Temperature'
bioracle_layers <- get_bio_oracle_v2(period = "Present",
                                     depth = "Surface",
                                     variable = "Temperature",
                                     save_path = ".")

# Download data for the 2050 period, RCP85, and the variable 'Chlorophyll'
bioracle_layers <- get_bio_oracle_v2(period = "2050",
                                     depth = "Surface",
                                     rcp = "RCP85",
                                     variable = "Temperature",
                                     format = "tif",
                                     save_path = ".")
}

}
