% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa.R
\name{get_chelsa}
\alias{get_chelsa}
\title{get_chelsa: Get bioclimatic layers from CHELSA}
\usage{
get_chelsa(
  period,
  model = NULL,
  rcp = NULL,
  sv_dir = getwd(),
  load2r = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{period}{Time period. Posible values are: "current","2041-2060","2061-2080"}

\item{model}{Circulation model. Posible values are:'ACCESS1-0',
'bcc-csm1-1', 'BNU-ESM', 'CanESM2','CCSM4','CESM1-BGC','CESM1-CAM5',
'CMCC-CM','CMCC-CMS','CNRM-CM5','CSIRO-Mk3-6-0','FGOALS-g2',
'FIO-ESM',GFDL-CM3','GFDL-ESM2G','GFDL-ESM2M','GISS-E2-H',
'GISS-E2-H-CC','GISS-E2-R','GISS-E2-R-CC','HadGEM2-AO',
'HadGEM2-CC','inmcm4','IPSL-CM5A-LR','IPSL-CM5A-MR','MIROC-ESM',
'MIROC-ESM-CHEM','MIROC5','MPI-ESM-LR','MPI-ESM-MR','MRI-CGCM3',
'MRI-ESM1','NorESM1-M'}

\item{rcp}{Representative Concentration Pathway. Posible values are "rcp26","rcp45","rcp85".}

\item{sv_dir}{Path to the directory where the layers will be saved. Default is the working directory of R session.}

\item{load2r}{Logical. Load layers into R?}

\item{parallel}{Download layers in parallel.}
}
\description{
Get bioclimatic layers from CHELSA for present and future scenarios
}
\details{
For more details visit \url{http://chelsa-climate.org/}
}
\examples{
\dontrun{
chelsa_biocurrent <- get_chelsa(period = "current",
                                sv_dir = "~/Desktop")
chelsa_bioCCSM4 <- get_chelsa(period = "2041-2060",
                              model="CCSM4",
                              rcp="rcp85",
                              sv_dir = "~/Desktop")
}
}
\references{
Karger, D.N., Conrad, O., Bohner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W., Zimmermann, N.E., Linder, H.P. & Kessler, M. (2017) Climatologies at high resolution for the earth's land surface areas. Scientific Data 4, 170122.
}
\seealso{
\code{\link[ntbox]{get_envirem_elev}}, \code{\link[ntbox]{get_envirem_clim}},\code{\link[ntbox]{get_bio_oracle}}.
}
