% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipselect.R
\name{inEllipsoid}
\alias{inEllipsoid}
\title{ellipsoid_selection: Performs variable selection for ellipsoid models}
\usage{
inEllipsoid(centroid, eShape, env_data, level)
}
\arguments{
\item{centroid}{A numeric vector of centroids for each environmental variable}

\item{eShape}{Shape matrix of the ellipsoid (can be a covariance matrix or a minimum volume ellipsoid).}

\item{env_data}{A data frame with the environmental training data.}

\item{level}{Proportion of points to be included in the ellipsoids. This paramter is equivalent to the error (E) proposed by Peterson et al. (2008).}
}
\description{
Performs variable selection for ellipsoid models according to omission rates in the environmental space.
}
