% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc}
\alias{nc}
\title{nc: Function to check the number of available cores}
\usage{
nc(ncores)
}
\arguments{
\item{ncores}{Number of cores for the parallel process.}
}
\value{
Returns an integer representing the number of cores that will be used to run a parallel process.
}
\description{
nc is a helper function to check if the number of cores required to run a parallel process is less or equal to the total number of cores of the system.
}
\details{
If ncores is bigger than the system's number of cores the function will return the system's number of cores. The functions that use this helper are \code{\link{mop}} and \code{\link{pROC}}.
}
\examples{
# Print the number of cores
print(nc(ncores=8))
}
