% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntb_mess.R
\name{ntb_mess}
\alias{ntb_mess}
\title{ntb_mess: Multivariate Environmental Similarity Surfaces (MESS)}
\usage{
ntb_mess(M_stack, G_stack)
}
\arguments{
\item{M_stack}{a RasterStack or a Matrix of variables representing the calibration area (M area in ENM context). If M_stack is matrix it should contain the values of environmental variables as get it from \code{\link[raster]{getValues}} function.}

\item{G_stack}{a RasterStack or a Matrix of variables representing areas or scenarios to which models will be transferred. If G_stack is matrix it should contain the values of environmental variables as get it from \code{\link[raster]{getValues}} function.}
}
\value{
A rasterLayer with MESS values
}
\description{
ntb_mess calculates multivariate environmental similarity surfaces as described by Elith et al., (2010) and optimized from the \code{\link[dismo]{mess}} function of the \code{\link[dismo]{dismo-package}}.
}
\examples{
m_stack <- raster::stack(list.files(system.file("extdata",
                                                package = "ntbox"),
                                    pattern = "M_layers.tif$",
                                    full.names = TRUE))
g_stack <- raster::stack(list.files(system.file("extdata",
                                                package = "ntbox"),
                                    pattern = "G_layers.tif$",
                                    full.names = TRUE))

messVals <- ntbox::ntb_mess(M_stack = m_stack,
                            G_stack = g_stack)
raster::plot(messVals)
}
\references{
Elith J., M. Kearney M., and S. Phillips, 2010. The art of modelling range-shifting species. Methods in Ecology and Evolution 1:330-342.
}
