% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_history.R
\name{occs_history}
\alias{occs_history}
\title{Function to visualize GIBIF data using googleVis}
\usage{
occs_history(data)
}
\arguments{
\item{data}{A data.frame with GBIF data}
}
\value{
Returns a list with googleVis motin chart and a data.frame
}
\description{
Visualize history of GBIF records
}

