% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_history.R
\name{occs_history_gbif}
\alias{occs_history_gbif}
\title{Function to visualize GIBIF data using googleVis}
\usage{
occs_history_gbif(gbif_data)
}
\arguments{
\item{gbif_data}{A data.frame with GBIF data}
}
\value{
Returns a list with googleVis motin chart and a data.frame
}
\description{
Visualize history of GBIF records
}
