% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omission_rate.R
\name{omission_rate}
\alias{omission_rate}
\title{Compute omission rate of a model}
\usage{
omission_rate(model, threshold, occ_train, occ_test)
}
\arguments{
\item{model}{Ecological niche model in raster format.}

\item{threshold}{A numeric vector with the thresholds that will be used to compute the omission rates.}

\item{occ_train}{Train data with three columns: "sp_name", "longitude" and "latitude"}

\item{occ_test}{Test data with three columns: "sp_name", "longitude" and "latitude"}
}
\value{
Returns a data.frame with the threshold value and the omission rate.
}
\description{
Computes the omission rate value of a model given a threshold
}
\examples{
\dontrun{
# Occurrence data for the giant hummingbird (Patagona gigas)
pg <- utils::read.csv(system.file("extdata/p_gigas.csv",
                                  package = "ntbox"))
# Split occs in train and test
pgL <- base::split(pg,pg$type)
pg_train <- pgL$train[,-4]
pg_test <- pgL$test[,-4]
mod_path <- file.path(system.file("extdata",
                            package = "ntbox"),
                     "p_gigas.tif")
model <- raster::raster(mod_path)
thresholds <- seq(0.01, 0.1,by = 0.03)
omrs <- omission_rate(model = model,
                      threshold = thresholds,
                      occ_train = pg_train,
                      occ_test = pg_test)

print(omrs)
}
}
