% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pROC.R
\name{pROC}
\alias{pROC}
\title{Partial ROC calculation for Niche Models}
\usage{
pROC(continuos_mod, test_data, n_iter = 1000, E_percent = 5,
  boost_percent = 50, parallel = TRUE)
}
\arguments{
\item{continuos_mod}{a RasterLayer of the ecological niche model to be evaluated.}

\item{test_data}{a numerical matrix containing coordinates of the occurrences used to test
the ecological niche model to be evaluated; columns must be: longitude and latitude.}

\item{n_iter}{(numeric) number of bootstrap iterations to be performed;
default = 1000.}

\item{E_percent}{(numeric) value from 0 to 100 that will be used as threshold (E);
default = 5.}

\item{boost_percent}{(numeric) value from 0 to 100 representing the percent of testing data
to be used for performing the bootstrap process for calculating the partial ROC;
default = 50.}

\item{parallel}{Logical to specify if the computation will be done in parallel. default=TRUE.}
}
\value{
A data.frame containing the AUC values and AUC ratios calculated for each iteration.
}
\description{
pROC applies partial ROC tests to continues niche models.
}
\details{
Partial ROC is calculated following Peterson et al.
(2008; \url{http://dx.doi.org/10.1016/j.ecolmodel.2007.11.008}). This function is a modification
of the \code{\link[ENMGadgets]{PartialROC}} funcion, available at \url{https://github.com/narayanibarve/ENMGadgets}.
}
