% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nichetoolbox.R
\name{run_nichetoolbox}
\alias{run_nichetoolbox}
\title{Function to run NicheToolBox shinyApp}
\usage{
run_nichetoolbox()
}
\description{
This function initiates NicheToolBox GUI
}
\examples{
# Run nichetoolbox package
# run_nichetoobox()
}

