% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ntbox.R
\name{run_ntbox}
\alias{run_ntbox}
\title{Function to run NicheToolBox shinyApp}
\usage{
run_ntbox()
}
\description{
This function initiates NicheToolBox GUI
}
\examples{
# Run nichetoolbox package
# run_nichetoobox()
}

