% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_envbg.R
\name{sample_envbg}
\alias{sample_envbg}
\title{Generate random environmental background data}
\usage{
sample_envbg(
  envlayers,
  nbg,
  nprop = NULL,
  coordinates = FALSE,
  cellIDs = FALSE,
  rseed = NULL,
  ncores = 4
)
}
\arguments{
\item{envlayers}{A raster stack or brick.}

\item{nbg}{Number of points for the background data}

\item{nprop}{Proportion of environmental data to be sampled. Default NULL}

\item{coordinates}{Logical. If TRUE cell coordinates will be returned}

\item{cellIDs}{Logical. If TRUE cell IDs will be returned}

\item{rseed}{Random seed number. Default NULL}

\item{ncores}{Number of workers to run the parallel process.}
}
\description{
Generate environmental background data is a function similar
to sampleRandom function of the raster package but optimized for Ecological
niche modeling.
}
\examples{
\dontrun{
wcpath <- list.files(system.file("extdata/bios",
                                 package = "ntbox"),
                     pattern = ".tif$",
                     full.names = TRUE)

envlayers <- raster::stack(wcpath)
vals <- sample_envbg(envlayers,nbg = 3583)
# Using a proportion of data
vals <- sample_envbg(envlayers,nprop = 0.20)
}
}
