% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_gbif_data.R
\name{search_gbif_data}
\alias{search_gbif_data}
\title{Function to search and download gbif data}
\usage{
search_gbif_data(
  genus,
  species,
  occlim = 10000,
  writeFile = FALSE,
  leafletplot = FALSE,
  showClusters = FALSE,
  ...
)
}
\arguments{
\item{genus}{Species genus}

\item{species}{Species name}

\item{occlim}{Occurrence data search limit.}

\item{writeFile}{Write gibif data into a csv}

\item{leafletplot}{Logical, if TRUE the records will be plotted on a leaflet map.}

\item{showClusters}{Logical. Display geographic cluster on the leaflet map.}

\item{...}{Pass arguments to the function  \code{\link[spocc]{occ}}  from spocc}
}
\value{
Returns a data.frame with coordinate data from species
}
\description{
Searchs gbif data using the function \code{\link[spocc]{occ}} from the spocc package
}
\details{
Note that this function is for exploration proposes.
If you want to use GBIF data in publications you should generate and cite
DOI information according to \href{https://www.gbif.org/citation-guidelines}{this}
}
\examples{
\dontrun{
# Species genus
genus <- "ambystoma"
# Species name
species <- "tigrinum"
# GBIF search
ambystoma_tigrinum <- ntbox::search_gbif_data(genus,species,
                                       occlim=100,
                                       writeFile=FALSE)
head(ambystoma_tigrinum[,1:5])
}

}
