% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca}
\alias{spca}
\title{spca: Principal Component Analysis for Spatial Data}
\usage{
spca(
  layers_stack,
  layers_to_proj = NULL,
  pca_obj = NULL,
  sv_dir = NULL,
  layers_format = ".asc",
  sv_proj_dir = NULL
)
}
\arguments{
\item{layers_stack}{A RasterStack of envrionmental variables}

\item{layers_to_proj}{A RasterStack of the envrionmental variables that will be projected (default NULL). If provided
the function will project the PCA for those layers by using the PCA object computed for the `layers_stack`. It can also use the PCA object
stored in `pca_obj`.}

\item{pca_obj}{An object of class \code{\link[stats]{prcomp}} (default NULL). Usefull when the user already
has done the PCa for `layers_stack`
and just wants to project the PCA on the `layers_to_proj` object.}

\item{sv_dir}{A directory where the PCs will be saved. If NULL the PCs will not be written.}

\item{layers_format}{A raster format for writing PCA results (see \code{\link[raster]{writeFormats}}). Default = ".asc"}

\item{sv_proj_dir}{A directory where the PCs projection will be saved. If NULL the PCs will be written inside sv_dir.}
}
\value{
A list containing either the raster stack with Pricipal Components of `layers_stack` or `layers_to_proj`,
a barplot of the cumulative and explained variance of each compoent of `layers_stack` and a \code{\link[stats]{prcomp}} object(`pca_obj`).
}
\description{
Performs PCA for a stack of raster layers.
}
\details{
spca uses the function \code{\link[stats]{prcomp}} of the `stats` package. If `sv_dir` is provided  the PCs
and `pca_obj` will be stored on it. The names of the layers in `layers_to_proj` need to be named
with exactly the same names of those either in`layers_stack` or the `pca_obj`.
}
\examples{
# -------------------------------------------------------------------
# A PCA on layers_stack  without saving.
layers_stack <- raster::stack(list.files(system.file("extdata",
                                    package = "ntbox"),
                                    pattern = "M_layers.tif$",
                                    full.names = TRUE))
pcs <- spca(layers_stack, sv_dir=NULL,layers_format=NULL)
raster::plot(pcs$pc_layers)

\dontrun{
# -------------------------------------------------------------------
# PCA projection  without saving.

layers_to_proj <- raster::stack(list.files(system.file("extdata",
                                                      package = "ntbox"),
                                          pattern = "G_layers.tif$",
                                          full.names = TRUE))
pcs_with_proj <- spca(layers_stack =layers_stack,
                      layers_to_proj = layers_to_proj,
                      pca_obj = NULL,
                      sv_dir=NULL,layers_format=NULL)
# Barplot of the Cumulative and explained variance in each component,
print(pcs_with_proj$pca_plot)
}
# -------------------------------------------------------------------
# PCA projection and saving.
\dontrun{
pcs_with_proj_sv <- spca(layers_stack =layers_stack,
                         layers_to_proj = layers_to_proj,
                         pca_obj = NULL,
                         sv_dir=".",layers_format=".asc")

}
# -------------------------------------------------------------------
# PCA projection, saving and using the pca_obj.
\dontrun{

pca_obj <- base::readRDS(list.files(system.file("extdata",
                                                package = "ntbox"),
                                                pattern = ".rds$",
                                                full.names = TRUE))

pcs_with_proj_sv <- spca(layers_stack =NULL,
                         layers_to_proj = layers_to_proj,
                         pca_obj = pca_obj,
                         sv_dir=".",layers_format=".tif")

raster::plot(pcs_with_proj_sv$pcs_layers_projection)
}
}
