% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_finder.R
\name{correlation_finder}
\alias{correlation_finder}
\title{Function to find out strong correlations in a correlation matrix}
\usage{
correlation_finder(cor_mat, threshold, verbose = TRUE)
}
\arguments{
\item{cor_mat}{A correlation matrix.}

\item{threshold}{Threshold valuefrom which it is considered that the correlation is high.}

\item{verbose}{Verbose output.}
}
\value{
Returns a vector with variable names that can summarize the information.
}
\description{
The function finds out which variables have strong
correlations according to a correlation threshold. The output
returns a list of variables names that can summarize the information
and removes the variables that are redundant.
}
\examples{
\dontrun{
environ_data <- read.csv(system.file("extdata",
                                     "nichekmeansCluster.csv",
                                      package = "ntbox"))[,-(1:3)]
environ_cors <- cor(environ_data)
correlation_finder(cor_mat = environ_cors ,threshold = 0.7,verbose = TRUE)
}
}
