% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_cluster_plot_3d.R
\name{ellipsoid_cluster_plot_3d}
\alias{ellipsoid_cluster_plot_3d}
\title{Function to plot clusters}
\usage{
ellipsoid_cluster_plot_3d(niche_data, cluster_ids, x, y, z, ellips, alpha,
  grupos, vgrupo, cex1 = 0.25, level)
}
\arguments{
\item{niche_data}{A data.frame or matrix containing niche variables}

\item{cluster_ids}{Cluster ids}

\item{x}{variable on x-axes}

\item{y}{variable on y-axes}

\item{z}{variable on z-axes}

\item{ellips}{Ellipsoid plot of each cluster}

\item{alpha}{Transparency level of ellipsoid plots}

\item{grupos}{Logical. Show ids of a geographic grouping variable}

\item{vgrupo}{A vector containing the levels of the  geographic grouping variable}

\item{cex1}{Point size}

\item{level}{proportion of points inside the ellipsoid}
}
\value{
Returns a data.frame with coordinate data from species
}
\description{
3D plot of clusters
}

