% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoidfit.R
\name{ellipsoidfit}
\alias{ellipsoidfit}
\title{Function fit an ellipsoid model}
\usage{
ellipsoidfit(data, centroid, covar, level = 0.95, threshold = 0.05,
  plot = T, size, xlab1 = "niche var 1", ylab1 = "niche var 2",
  zlab1 = "Suitability", ...)
}
\arguments{
\item{data}{A RasterStack or RasterBrick objet of the niche varibles.}

\item{centroid}{A vector with the values of the centers of the ellipsoid (see \code{\link[hdm]{cov_center}}).}

\item{covar}{The shape matrix (covariance) of the ellipoid (see \code{\link[hdm]{cov_center}}).}

\item{level}{The proportion of points  to be included inside the ellipsoid}

\item{threshold}{Threshold value for the suitabilities to be cosidered as 0,
by default all suiabilities <0.05 are cosidered as zeros}

\item{plot}{Logical If True a plot of niche will be shown.}

\item{size}{The size of the points of the niche plot.}

\item{xlab1}{For x label for 2-dimensional histogram}

\item{ylab1}{For y label for 2-dimensional histogram}

\item{zlab1}{For z label for 2-dimensional histogram}

\item{...}{Arguments passed to \code{\link[rgl]{plot3d}} function from rgl}
}
\value{
Returns a list containing a data.frame with the suitability values; a suitability raster;
a data.frame with the mahalanobis and euclidean distances to the centroid.
}
\description{
Function fit an ellipsoid model using the shape matrix (covariance matrix)
of the niche variables.
}
\examples{
## Load niche data
# d_cardon <-  read.csv(system.file("extdata", "cardon_virtual.csv", package = "nichetoolbox"))
## Compute the centroid and shape (covariance matrix) of the ellipsoid model.
# covar_centroid <- cov_center(d_cardon,mve=TRUE,level=0.99,vars=c(3,4,5))
## RasterStack with the niche variables
# nicheStack <- stack(list.files(system.file("extdata",
#                   package = "nichetoolbox"),
#                   pattern = ".asc$",full.names = TRUE))
# Fitting the ellipsoid model
# ellipsoidMod <- ellipsoidfit(nicheStack,
#                          covar_centroid$centroid,
#                          covar_centroid$covariance,
#                          level=0.95,threshold=0.05,plot=TRUE,size=3)
# plot(ellipsoidMod$suitRaster)
}
