% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_history.R
\name{occs_history}
\alias{occs_history}
\title{Function to visualize GIBIF data using googleVis}
\usage{
occs_history(gbif_data)
}
\arguments{
\item{gbif_data}{A data.frame with GBIF data}
}
\value{
Returns a list with googleVis motion chart and a data.frame
}
\description{
Visualize a time series of the history of GBIF records.
}
\details{
Creates a motionchart of the time series of the occurrence data. It uses the function \code{\link[googleVis]{gvisMotionChart}} from \code{\link[googleVis]{googleVis-package}}.
}
\examples{
\dontrun{
ambystoma_tigrinum <- searh_gbif_data(genus = "Ambystoma",
                                      "tigrinum",occlim=1000,
                                      writeFile=FALSE)

atigrinum_history <- occs_history(ambystoma_tigrinum)
plot(atigrinum_history$mot)
}
}
