% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlayers_ntb.R
\name{rlayers_ntb}
\alias{rlayers_ntb}
\title{Function to read raster layers with the same extent and resolution in all `raster` acepted formats}
\usage{
rlayers_ntb(layers_path)
}
\arguments{
\item{layers_path}{Path to the folder conteining raster layers. All layers in the folder have to have the same extent and resoltion}
}
\value{
Returns a raster stack of raster layers with the same extent and resolution
}
\description{
Read raster layers that that have the same extent and resolution,
all  \code{\link[raster]{raster-package}} formats are accepted (see \code{\link[raster]{writeFormats}})
}
\examples{
\dontrun{
## RasterStack with the niche variables
nicheStack <- rlayers_ntb(list.files(system.file("extdata",
                                     package = "nichetoolbox"),
                                     pattern = ".asc$",full.names = TRUE))
}
}
